[Rename] 3단계 자동차 경주 -> racing/basic  폴더로 이동
